Adds monitoring for lenovo machines that use the xclarity controller.
The idrac exporter works out of the box for the lenovo machines. The
ipmi exporter required that I explicitly specify cihper suite 17 as the
lenovo machines have some security policy that prevent them from using
older ciphers. The dell idracs also support this cipher suite as far as
I can tell.

For the snmp exporter to work we'll need to do the dance with lenovo
specific mibs.