Create a prometheus recording rule to precompute a new field
From the output of idrac_system_machine_info, we want to use the "sku" field
for dell machines since that corresponds to the service tag but use "serial"
field for all other manufacturers. With this rule we compute a new
metric idrac_system_machine_info_fixed that sets a new label unique_id
based on the manufacturer label. We then use unique_id in the alerts we
raise.